requirements.txt
appnope==0.1.0
coreapi==2.3.1
coreschema==0.0.4
decorator==4.0.9
Django==1.11.5
django-filter==1.0.4
django-webpack-loader==0.5.0
djangorestframework==3.6.4
gnureadline==6.3.3
hvac==0.2.16
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.7.3.1
ptyprocess==0.5.1
Pygments==2.2.0
pytz==2017.2
requests==2.12.4
simplegeneric==0.8.1
traitlets==4.1.0
uritemplate==3.0.0
virtualenv==13.1.2
